Updated OrderSession
Before when the program was validating for a certain box type it would get all the cases in which it should return false and if none of them were met it would return true. That seemed a bit backwards, I don't really know why I decided to do it like it, but now it gets all the conditions for which the box validation would return true and if not all of them are met it should return false.